feat: add CryptographyService abstraction
