+++ packages/dart/cryptography/lib/src/cryptography/cryptography_abstract.dart
@@ -1,0 +1,17 @@
+import 'package:affinidi_tdk_cryptography/affinidi_tdk_cryptography.dart';
+
+abstract class CryptographyServiceAbstract
+    implements CryptographyServiceInterface {
+  static CryptographyServiceAbstract? _instance;
+
+  static CryptographyServiceAbstract get instance {
+    if (_instance == null) {
+      throw StateError('CryptographyServiceAbstract instance not set.');
+    }
+    return _instance!;
+  }
+
+  static set instance(CryptographyServiceAbstract instance) {
+    _instance = instance;
+  }
+}
